+++ res/texturepacks/default/info.lua
@@ -9,11 +9,23 @@ return {
         }
     },
     font = {
-        source = 'imagefont.png',
-        glyphs = {
-            source = ' ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyzÄÖÜäöü0123456789.,:;!?-+/()[]%&"\'*=_<>ß^©|',
-            width = 8,
-            height = 16
+        width = 8,
+        height = 16,
+        charsets = {
+            {
+                -- LATIN BASIC
+                -- 0020-007F (Excluded: 007F)
+                --  !"# $%&' ()*+ ,-./ 0123 4567 89:; <=>? @ABC DEFG HIJK LMNO PQRS TUVW XYZ[ \]^_ `abc defg hijk lmno pqrs tuvw xyz{ |}
+                source = 'imagefont_latin_basic.png',
+                glyphs = [[ !"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\]^_`abcdefghijklmnopqrstuvwxyz{|}~]]
+            },
+            {
+                -- LATIN-1
+                -- 00A0-00FF (Excluded: 00A0, 00AD)
+                -- ¡¢£¤ ¥¦§¨ ©ª«¬ ®¯°± ²³´µ ¶·¸¹ º»¼½ ¾¿ÀÁ ÂÃÄÅ ÆÇÈÉ ÊËÌÍ ÎÏÐÑ ÒÓÔÕ Ö×ØÙ ÚÛÜÝ Þßàá âãäå æçèé êëìí îïðñ òóôõ ö÷øù úûüý þÿ
+                source = 'imagefont_latin_1.png',
+                glyphs = [[¡¢£¤¥¦§¨©ª«¬®¯°±²³´µ¶·¸¹º»¼½¾¿ÀÁÂÃÄÅÆÇÈÉÊËÌÍÎÏÐÑÒÓÔÕÖ×ØÙÚÛÜÝÞßàáâãäåæçèéêëìíîïðñòóôõö÷øùúûüýþÿ]]
+            }
         }
     }
 }