+++ src/turnbased/states/PlanningState.lua
@@ -71,7 +71,7 @@ local function handleInputSelectionActions( input, inputStateHandler, character 
     return true
 end
 
-local function handleOtherActions( input, stateManager, inputStateHandler, factions, character )
+local function handleOtherActions( input, stateManager, inputStateHandler, factions, character, explosionManager, projectileManager )
     if input == 'next_character' then
         inputStateHandler:switch( 'movement' )
         factions:getFaction():nextCharacter()
@@ -90,7 +90,7 @@ local function handleOtherActions( input, stateManager, inputStateHandler, facti
 
     if input == 'open_inventory_screen' then
         character:enqueueAction( OpenInventory( character, character:getTile() ))
-        stateManager:push( 'execution', factions, character )
+        stateManager:push( 'execution', factions, character, explosionManager, projectileManager )
         return true
     elseif input == 'open_health_screen' then
         ScreenManager.push( 'playerInfo', character )
@@ -127,8 +127,10 @@ function PlanningState:initialize( stateManager )
     self.inputStateHandler:switch( 'movement' )
 end
 
-function PlanningState:enter( factions )
+function PlanningState:enter( factions, explosionManager, projectileManager )
     self.factions = factions
+    self.explosionManager = explosionManager
+    self.projectileManager = projectileManager
 end
 
 function PlanningState:update()
@@ -148,7 +150,7 @@ function PlanningState:input( input )
     end
 
     if handleCharacterActions( input, character ) then
-        self.stateManager:push( 'execution', self.factions, character )
+        self.stateManager:push( 'execution', self.factions, character, self.explosionManager, self.projectileManager )
         return
     end
 
@@ -156,7 +158,7 @@ function PlanningState:input( input )
         return
     end
 
-    if handleOtherActions( input, self.stateManager, self.inputStateHandler, self.factions, character ) then
+    if handleOtherActions( input, self.stateManager, self.inputStateHandler, self.factions, character, self.explosionManager, self.projectileManager ) then
         return
     end
 end
@@ -174,9 +176,9 @@ function PlanningState:selectTile( tile, button )
     end
 
     -- Request actions to execute.
-    local execute = self.inputStateHandler:getState():request( tile, character )
+    local execute = self.inputStateHandler:getState():request( tile, character, self.projectileManager )
     if execute then
-        self.stateManager:push( 'execution', self.factions, character )
+        self.stateManager:push( 'execution', self.factions, character, self.explosionManager, self.projectileManager )
     end
 end